Move components to nav folder
