+++ src/pages/course/[courseID]/index.tsx
@@ -13,9 +13,9 @@ import { isFinite, toNumber } from "lodash";
 import { useRouter } from "next/router";
 
 import ChakraMarkdown from "@/components/ChakraMarkdown";
-import { Hero } from "@/components/Hero";
 import Page from "@/components/Page";
 import KnowledgeCard from "@/components/feed_cards/KnowledgeCard";
+import { Hero } from "@/components/nav/Hero";
 import CourseModel from "@/lib/db/models/Course.model";
 import { HTTPBadRequest } from "@/lib/errors/HTTPErrors";
 import { useCallback } from "react";
@@ -60,8 +60,7 @@ const CoursePage = ({
           title={title}
           module={<Text color='text'>{description}</Text>}
         />
-      }
-    >
+      }>
       {knowledge.map((k, index) => (
         <KnowledgeCard
           key={k.id}
@@ -69,8 +68,7 @@ const CoursePage = ({
           footer={
             <Text color='gray.500'>{`${k.questionIDs.length} questions to study`}</Text>
           }
-          onClick={onClickFactory(k.id)}
-        >
+          onClick={onClickFactory(k.id)}>
           <ChakraMarkdown>{k.text}</ChakraMarkdown>
         </KnowledgeCard>
       ))}
