+++ src/pages/course/questions.tsx
@@ -6,10 +6,10 @@ import { Text } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 
 import Feed from "@/components/Feed";
-import { Hero } from "@/components/Hero";
 import { Main } from "@/components/Main";
 import Page from "@/components/Page";
 import UserInput from "@/components/UserInput";
+import { Hero } from "@/components/nav/Hero";
 import QuestionModel, { TQuestionData } from "@/lib/db/models/Question.model";
 
 type Props = {
@@ -26,8 +26,7 @@ const QuestionsPage = ({ questions = [] }: Props) => {
           module={<Text color='text'>Economics - {router.query.courseID}</Text>}
         />
       }
-      footer={<UserInput />}
-    >
+      footer={<UserInput />}>
       <Main>
         <Feed questions={questions} />
       </Main>
